requirements.txt
openai==1.55.0
gspread==6.1.4
google-auth==2.36.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1